==> scripts/04_define_connotation_groups.R <==

# Load packages
library(tidyverse)
library(readr)

# Load data
us <- read_csv('output-data/us_names_with_popularity_and_connotations.csv')
uk <- read_csv('output-data/uk_names_with_popularity_and_connotations.csv')

# Next we define connotation groups. This was done through an LLM assisted process, where we iteratively fed the LLM the list of unassigned connotations and asked it if any belonged in any of these categories. Boundaries between concepts are subjective, and this is a limitation of this approach which ultimately relies in part on human judgement (is, for instance, "graceful" sufficiently related to beauty to belong to that category?) 
connotation_groups <- list(
  intelligence = unlist(c("intelligence", 
                          c(strsplit(c('wisdom, wise, knowledge, intelligence, knowledgeable, king of wisdom, clever, sacred knowledge, wise protector, smart, servant of the wise, vedic knowledge, cleverness, knowledgeable, goddess of wisdom, symbol of wisdom, elf-wise, lover of wisdom, wise legislator, thirst for knowledge, scholarly knowledge, keeper of knowledge, wise leader, poetry and wisdom, divine knowledge, ancient wisdom, little old wise one, wise woman, street-smart or cunning, wise child, esoteric wisdom, smart devices, wise teacher, wise advisor, spiritual wisdom, wise counselor, education, academic, scholarly, education, academic excellence, scholarly'), ', ')),
                          strsplit(c("academic, academic excellence, achievement, analytical, academic, academic excellence, achievement, analytical"), ", "), c(
                            "brilliance",
                            "creativity",
                            "creative",
                            "curiosity",
                            "detective",
                            "eloquent",
                            "enlightened",
                            "enlightenment",
                            "excellence",
                            "sophistication",
                            "sophisticated",
                            "genius",
                            "gifted",
                            "guidance",
                            "illumination",
                            "innovation",
                            "innovative",
                            "insight",
                            "insightful",
                            "intellect",
                            "intellectual",
                            "intelligent",
                            "inventor",
                            "philosopher",
                            "philosophy",
                            "scholar",
                            "scientific",
                            "tech-savvy",
                            "thoughtful",
                            "understanding",
                            "bright",
                            "visionary",
                            "university",
                            "imaginative",
                            "sophisticated",
                            "skilled",
                            "expert"
                          )
  )),
  beauty = c(unlist(c("beauty", 
                    strsplit(c("elegance, graceful, gracefulness, grace, beautiful, beauty, beautiful friend, beautiful moon, attractive, french elegance, spring beauty, elegance, beautiful night child, summer beauty, elegance, elegance, meliza connotes elegance, beautiful blossom, beautiful flower, roman goddess of love and beauty, beauty, full of grace, goddess of beauty, elegance, dark beauty, hawaiian (meaning 'beautiful'), natural beauty, beautiful eyes, beautiful, attractiveness, morning beauty, beauty, slavic goddess of love and beauty, hawaiian (beautiful), beauty, divine beauty, lovely, vintage elegance, beautiful harbor, beautiful bloom, coastal beauty, beautiful child, beautiful blessing, beautiful smile, graceful, beautiful woman, heavenly beauty, elegance, rare beauty, elegance and grace, lunar beauty, classic beauty, exotic beauty, beauty and tranquility, delicate beauty, gentle beauty, serene beauty, celestial beauty, harmonious beauty, mediterranean beauty, classical beauty, beautiful future, classic elegance, remote beauty, strength and beauty, black beauty, beautiful generation, tuscan beauty, timeless beauty,  everlasting beauty, blessed beauty, enduring beauty, earthly beauty,  elegant beauty"), ', '))),
             "allure",
             "alluring",
             "cute",
             "elegant",
             "enchanting",
             "ethereal",
             "glamorous",
             "glamour",
             "gracious",
             "handsome",
             "heavenly"),
  strength = c(unlist(c("strength", 
                      strsplit(c("strength, strong, noble strength, power, first power, economic powerhouse, goddess of power, strong leader, short and strong, powerful, strength,  mighty, strong, empowerment and strength, strong woman, immense power, dark and powerful, strongman, powerful ruler, strong character, empowered, strong female leader, battle-mighty, strong warrior, strength from god, strong businesswoman, strong and independent woman, symbol of power and authority, strong and brave, strong and rugged, feminine strength, force, strong protector, strong personality, strong leadership, strong and reliable,  strong faith, symbol of strength, powerful warrior, power and authority, rugged or strong, strong or masculine, mighty one, strong and masculine, strength and resilience, large or powerful, strength and loyalty, mighty warrior, strong and supportive partner, strength and power, powerful woman, powerful being, powerful and destructive, strong or steadfast, strengthener, strong ally, strong and heroic"), ', '))), unlist(strsplit(c("strength, strong, mighty, powerful"), ", ")),
               unlist(c(
                 "strength",
                 strsplit(c("strength, strong, powerful, powerful leader, strong character"), ", ")
               )),
               "strength"
               ), 
  
  
  wealth = c(unlist(c("wealth", 
                    strsplit(c("luxury, prosperity, prosperous, rich, wealthy, richness, prosperous, rich history, rags to riches, historical richness, affluent, rich, wealthy guardian, wealth and luxury, associated with luck and prosperity, prosperous companion, prosperous child"), ', '))),
  
             "abundance",
             "affluence",
             "fortunate",
             "generosity",
             "generous",
             "good fortune",
             "fortune",
             "prestige",
             "prestigious",
             "exclusivity",
             "precious",
             "valuable",
             "luxurious"
             ),
  
  
  love = unlist(c('love', 
                  strsplit(c("love, compassion, romance, beloved, affection, city of love, affectionate, goddess of love, divine love, buddhist deity of compassion, compassionate, beloved by god, beloved of god, beloved of the lord, roman goddess of love and beauty, love of god, lover, affectionate nickname, gift of love, slavic goddess of love and beauty, passion fruit, radha (a beloved consort of lord krishna in hindu mythology), worthy of love, lovely, passionate, loved, nature-lover, tragic love, beloved child, loveliness, affectionate term, god's affection, unrequited love, moon lover, god of love, god's love, adoration, nature lover, brotherly love, self-love, loved one, beloved companion, steadfast love, passionate love, peace and love, love interest, love, affection, passion, romance, beloved, compassion, beloved by god, beloved of god, affectionate, passion fruit, affectionate nickname"), ', '),
                  
                  
                  "love",
                  "affection",
                  "passion",
                  "romance",
                  "beloved",
                  "compassion",
                  "passion fruit",
                  "affectionate",
                  "affectionate nickname",
                  "lover",
                  "loving",
                  "loveliness",
                  "lovely",
                  "devotion",
                  "devoted",
                  "devotion",
                  "devout",
                  "loved",
                  "loved one",
                  "beloved companion",
                  "soul",
                  "soulful",
                  "loyal",
                  "loyalty",
                  "friend",
                  "friendly",
                  "friendliness",
                  "kind",
                  "kindness",
                  "kind-hearted",
                  "caring",
                  "nurturing",
                  "romantic",
                  "benevolence",
                  "companion"
                  )),
  
  
  
  
  joy = unlist(c('joy', 
                 strsplit(c("joyful, happiness, joy, bliss, delight, delightful, divine joy, joyful, blissful, divine happiness, deep happiness, joyful song, she brings happiness, spiritual bliss, father’s joy, bringer of happiness, joyous, delighted, one who brings joy, joy and happiness, pure joy, my joy, pure delight, spiritual joy, inner happiness, bringer of joy, eternal joy, celestial joy, eternal bliss, eternal happiness, merriment, god's joy, messenger of joy, child of joy, morning joy"), ', '),
                 "joy",
                 "happiness",
                 "bliss",
                 "delight",
                 "cheerful",
                 "cheerfulness",
                 "joyful",
                 "blissful",
                 "delight",
                 "joyful song",
                 "joyous",
                 "delighted",
                 "joy and happiness",
                 "pure joy",
                 "my joy",
                 "pure delight",
                 "spiritual joy",
                 "inner happiness",
                 "eternal joy",
                 "celestial joy",
                 "eternal bliss",
                 "eternal happiness",
                 "merriment",
                 "god's joy",
                 "messenger of joy",
                 "child of joy",
                 "morning joy",
                 "happy",
                 "festive",
                 "celebration",
                 "fun-loving",
                 "positivity",
                 "positive",
                 "hope",
                 "playful"
                 )),
  
  
  religious = unlist(c('religious', 
                       strsplit(c("divine song, sacred plant, sacred cow, sacred space, god's flock, beloved of god, dante's guide in the divine comedy, blessings, sacred experience, holy city, gods, helper of god, sacred tree, power of god, divine comedy, spiritual strength, god is my light, god-consciousness, spiritual bliss, goddess of the hunt, roman god of agriculture, spiritual retreats, divine elixir, roman goddess of agriculture, divine approval, god leads, divine provision, spiritual dedication, god will add, storm god, assyrian god, sacred or holy place, divine mother, divine, spiritual ascension, god of shiva, god’s mercy, goddess of war, celtic goddess, god's work, divine timing, supreme deity, goddess of fertility, deity, aztec rain god, born of the gods, divine connection, god's aid, divine unification, spiritual, god is good, spiritual sanctuary, worshipper, close to god, egyptian god, ancient deity, god is great, norse god odin, spiritual leader, divine presence, christian faith, sun god, messenger of the gods, worshipped, god of my people, ancient religion, beautiful blessing, god's honor, who is like god, goddess of sight, divine gift, 5 spiritual, gift from god, spiritual chant, hawaiian god of the sea, spiritual initiation, goddess of mist, divine guide, divine authority, sacred verses, prayer, god's treasure, spiritual transformation, venom of god, 2 blessings, warrior deity, spiritual song, god has helped, divine beauty, sumerian god of water, roman goddess, islamic holy city, spiritual joy, divine favor, sacred, buddhist deity of compassion, shinto religion, divine intervention, king of the gods, egyptian goddess, goddess sita, divine knowledge, warrior god, divine forgiveness, armenian goddess, god's power, divine nectar, god's love, divine hymns, islamic holy book, divine grace, worship, god of thunder and rain, spiritual awareness, holy, god's heritage, divine or heavenly, goddess ganga, god of agriculture, falcon-headed deity, godly, divine pleasure, goddess of the stars, chosen by god, spiritual refuge, faith, beloved of the lord, god's miracle, spiritual elevation, religious, spiritual, faith, divine, blessed, sacred, god, goddess, worship, sacred space, divine blessing, sacred cow"), ', '),
                       strsplit(c("god's glory, god's light, spiritual leader, spiritual elevation, spiritual refuge, god of the dead, divine fruit, son of god, spiritual journey, goddess of love and beauty, god of strength, god of war, spiritual enlightenment, holy spirit, goddess of wisdom, protector of faith, god's blessing, sacred flame, god's jewel, god of life, spiritual destination, god’s gift, god-loving, goddess of beauty, divine protection, strength from god, spiritual struggle, loyal servant of god, sea god, goddess of death, divine progenitor, my god has answered, spiritual symbolism, faithful servant, divine support, my god, sacred site, god's family, god has given, 3 divine, spiritual teachings, god's chosen one, spiritual learning, protector of faith, sacred river ganges, god is gracious, spiritual wisdom, sumerian deity, mayan goddess, moon goddess, sacred hymn, favor of god, 2 spiritual leadership, singular worship, faithful, god hears, religion, god has answered, divine announcement, spiritual elevation, roman god, ganga (the sacred river), irish goddess of summer, goddess parvati, love of god, sacred bull, solar deity, spiritual connection, divine inspiration, divine offering, child of god, hawaiian goddess of fertility and childbirth, hindu god of rain and thunderstorms, divine table, god is with us, daughter of god, yama (hindu and buddhist deity), favor from god, godliness, spiritual practice, home of the gods, place of worship, 1 divine, warrior of god, goddess-like, spiritual fulfillment, goddess of the sea, divine will, sky god, favored by god, unfaithful, divine punishment, blessed by god, 3 divine gifts, 3 spiritual, god (in italian), divine happiness, god's witness, slavic goddess of love and beauty, friend of god, goddess of love, roman goddess of love and beauty, god is my father, sacred river, divine and majestic, divine incarnation, slavic origin meaning 'gift of god', god is with me, divine play, sea goddess, divine communication, sacred basil, strong faith, god-like, divine child, the top 5 connotations of the name allah are: god, ancient egyptian god of air, inari fox (kitsune associated with the deity), 3 spirituality, aztec goddess, lioness of god, divine healer, god's ornament, spiritual depth, mother goddess, worshipful, goddess of power, faithful companion, god of trade, god's promise"), ', '),
                       "religious",
                       "spiritual",
                       "faith",
                       "divine",
                       "blessed",
                       "sacred",
                       "prophet",
                       "messiah",
                       "gift of god",
                       "christian",
                       "christianity",
                       "christmas",
                       "prophetess",
                       "martyr",
                       "messiah",
                       "easter",
                       "virgin mary",
                       "god",
                       "goddess",
                       "worship",
                       "sacred space",
                       "divine blessing",
                       "sacred cow",
                       "divine messenger",
                       "divinity",
                       "worshipper",
                       "prayer",
                       "worshipped",
                       "god's work",
                       "worship",
                       "religious devotion",
                       "religious significance",
                       "faithfulness",
                       "piety",
                       "religious significance",
                       "sacred flame",
                       "sacred river ganges",
                       "faith",
                       "holy",
                       "religious devotion",
                       "holy spirit",
                       "sacred site",
                       "holy city",
                       "religious faith",
                       "spiritual leader",
                       "sacred place",
                       "holy spirit",
                       "sacred or holy place",
                       "god of shiva",
                       "goddess of war",
                       "sacred river",
                       "holy book",
                       "divine guidance",
                       "biblical",                  "biblical figure",           "biblical hero",             "biblical prophet",         
                       "biblical reference",        "biblical significance",
                       "saint",
                       "saintly",
                       "angelic",
                       "angel",
                       "prophetic",
                       "pious",
                       "savior",
                       "archangel",
                       "blessing"
                     )), 
  
  
  tradition = unlist(c('tradition', 
                       strsplit("heritage, traditional, cultural heritage, spiritual, family heritage, spirituality, tradition, spiritual leader, remigio connotes tradition, spiritual power, spiritual connection, spiritual awakening, traditionalism, spiritual guide, clan tradition, spiritual journey, spiritual entity, noble heritage, literary tradition, spiritual bliss, traditional craft, spiritual heritage, ritual, spiritual learning, spiritual initiation, spiritual or religious, spiritual strength, traditional values, spiritual guidance, someone with unconventional or non-traditional views, spiritual transformation, spiritual teachings, spiritual leadership, spiritual quest, traditionalist, spiritual elevation, spiritual joy, spiritual fulfillment, god's heritage, spiritual enlightenment, spiritual growth, religious tradition, spiritual ascension, spiritual teacher, spiritual retreat, rich oral tradition, spiritual insight, spiritual center, spiritual sanctuary, ancient heritage, spiritual practice, spiritual light, spiritual awareness, traditional roots, ancient traditions, traditional profession, literary heritage, spiritual philosophy, traditional remedy, artistic heritage, customs, traditional heritage, spiritual song, maritime heritage, traditional pottery, spiritual symbolism, traditional yet unique, spiritual reminder, spiritual wisdom, orphic tradition, spiritual struggle, spiritual dedication, spiritual refuge, spiritual destination, non-traditional, traditional name, spiritual chant, spiritual retreats, traditional culture, traditional crafts, traditional craftsmanship, traditional principles, traditional recipes, traditional ceremonies, ancient tradition, cultural tradition, tradition, heritage, culture, traditional, cultural heritage, cultural pride, heritage, origin, ethnicity, traditional values, heritage, origin, ethnicity, cultural heritage, heritage, heritage, origin, ethnicity", ', '),
                       "tradition",
                       "heritage",
                       "culture",
                       "traditional",
                       "cultural heritage",
                       "cultural pride",
                       "heritage",
                       "origin",
                       "ethnicity",
                       "traditional values",
                       "spiritual",
                       "family heritage",
                       "spiritual power",
                       "spiritual connection",
                       "spiritual awakening",
                       "traditionalism",
                       "spiritual guide",
                       "clan tradition",
                       "spiritual journey",
                       "spiritual entity",
                       "noble heritage",
                       "literary tradition",
                       "spiritual bliss",
                       "traditional craft",
                       "spiritual heritage",
                       "ritual",
                       "spiritual learning",
                       "spiritual initiation",
                       "spiritual or religious",
                       "spiritual strength",
                       "spiritual guidance",
                       "spiritual transformation",
                       "spiritual teachings",
                       "spiritual leadership",
                       "spiritual quest",
                       "traditionalist",
                       "spiritual elevation",
                       "spiritual joy",
                       "spiritual fulfillment",
                       "spiritual enlightenment",
                       "spiritual growth",
                       "religious tradition",
                       "spiritual ascension",
                       "spiritual teacher",
                       "spiritual retreat",
                       "rich oral tradition",
                       "spiritual insight",
                       "spiritual center",
                       "spiritual sanctuary",
                       "ancient heritage",
                       "spiritual practice",
                       "spiritual light",
                       "spiritual awareness",
                       "traditional roots",
                       "ancient traditions",
                       "traditional profession",
                       "literary heritage",
                       "spiritual philosophy",
                       "traditional remedy",
                       "artistic heritage",
                       "customs",
                       "traditional heritage",
                       "spiritual song",
                       "maritime heritage",
                       "traditional pottery",
                       "spiritual symbolism",
                       "traditional yet unique",
                       "spiritual reminder",
                       "spiritual wisdom",
                       "orphic tradition",
                       "spiritual struggle",
                       "spiritual dedication",
                       "spiritual refuge",
                       "spiritual destination",
                       "non-traditional",
                       "traditional name",
                       "spiritual chant",
                       "italian heritage",  
                       "spiritual retreats",
                       "traditional culture",
                       "traditional crafts",
                       "traditional craftsmanship",
                       "traditional principles",
                       "traditional recipes",
                       "traditional ceremonies",
                       "ancient tradition",
                       "cultural tradition",
                       "irish heritage",
                       "irish origin",
                       "old-fashioned",
                       "historical",
                       # Heritage / Cultural origins
                       "irish heritage", "italian origin", "french origin", "english origin", 
                       "scottish heritage", "russian origin", "spanish origin", "german origin", 
                       "germanic origin", "hawaiian origin", "slavic origin", "scandinavian", 
                       "scandinavian heritage", "native american heritage", "japanese origin", 
                       "african heritage", "welsh origin", "dutch origin",
                       
                       # Historical / Legacy aspects
                       "ancient", "historical", "historical significance", "medieval", 
                       "mythology", "mythological", "history", "legacy",
                       
                       # Aristocratic / Nobility references
                       "aristocratic", "royalty", "nobility", "noble",
                       
                       # Cultural / Traditional references
                       "cultural", "southern charm", "place name"
                       
                       ))
  )

# Define a function to add connotation columns to a dataset
add_connotation_columns <- function(df, connotation_groups) {
  for (group in names(connotation_groups)) {
    terms <- connotation_groups[[group]]
    
    # Dynamically create the column name and assign TRUE/FALSE
    df <- df %>%
      mutate(!!group := if_any(starts_with("connotation_"), ~ .x %in% terms))
  }
  return(df)
}

# Apply the function to both us and uk datasets
us <- add_connotation_columns(us, connotation_groups)
uk <- add_connotation_columns(uk, connotation_groups)

write_csv(uk, 'output-data/uk_names_with_popularity_and_connotations.csv')
write_csv(us, 'output-data/us_names_with_popularity_and_connotations.csv')
